Implemented config file includes.
I've implemented the previously stubbed out "_file" config expression.  The root
JSON object of the included file becomes the value of the field.  In addition to
searching the CWD and the config dir, the include will search directories listed
in the CAMLI_INCLUDE_PATH environment variable.  The include path is useful
because it gives us the ability to have default configs in git which can be
overridden by custom configs which are provided in another path directory which
gets searched first.

I ended up working on this because, as I was thinking about how to approach
supporting Google Storage, I realized that we would need to have secrets in
includes; this would be a pain when combined with the current configs and
runscripts in GIT.  So, one thing lead to another, and now this is done.

Change-Id: Ib256859326dfceefed90d0fadf83a76adffd538b