find the path to the executable of the running instance to restart camli properly
This change adds the functionality to find the path to
the currently running executable, so that camlistore
can be restarted without à priori knowledge of cwd.
It should work for linux, netBSD and freeBSD, and it
might work for openBSD.
The naive (os.Args[0]) approach is still used with
darwin.

Reference: https://codereview.appspot.com/6736069/

Change-Id: Icc9fc2b4e1b3542cfbe6c457c1fefbe762eff2f1