images: Decode now returns a Config with a Modified field
This allows to know after a call to images.Decode if the image
was actually rotated or flipped.
Without knowing this, when making a thumbnail, it could happen
(if there was no rescaling required) that an image that needed
being rotated/flipped would not get rotated, because
useBytesUnchanged would still be true.

Change-Id: Ifc5e1e1f5a8543e6754102e3b3a685b736ae8673