closure: changed the library setup, added the compiler
1) The code to get the closure library has been moved from
dev-server to misc/get_closure.pl
2) The svn checkout version for the closure library was bumped
to the latest: 2459
3) The closure library checkout now goes into tmp/closure-library/,
to keep tmp/ clean for other things we might want to add there
(like the compiler). So the path to the actual libraries is now
tmp/closure-library/closure/
4) The code to get the closure compiler build was added in
misc/get_closure.pl. The compiler goes to
tmp/closure-compiler/compiler.jar
5) The version of the closure compiler is the latest build as
of this time: 20121212

Change-Id: I7e03ad7d6fe440430c5727cf56a956efed360d43