blob: replacement for *blobref.BlobRef type.
Shorter name (blob.Ref instead of *blobref.BlobRef) and value
semantics instead of pointer.

Also, equality (==) works on it, so it can be used as a map key.

It's also more memory efficient, for the upcoming search server that
needs to store tons (or all) of your blobrefs in memory.

Change-Id: I494a2d72d31865211fa388fedf938ed25f99ca13