server/image: properly set image format.
The value of format was empty and causing cr2 and jpeg images to fall through
to the the default case, encoding thumbnails as PNG.

This change sets the default case as jpeg, and ensures that cr2 is encoded as
jpeg.  Png files now have their own explicit branch in the switch.

Tested by uploading a cr2, jpeg and png file to camlistore, then viewed the
mime type in the thumbnail response for each.

Change-Id: Ibe4400f220014b3fd72d9ae420c7d60c2f8cb33a