buildbot: Add more debugging.
This adds the error returned from exec to the log message, and
provides '(empty)' placeholders when no stdout or stderr is present
from the called binary.

This has the effect of changing the silent errors I was seeing on my
buildbot from:

"""
Error:
	:
"""

To:

"""
Error:

	exec: "devcam": executable file not found in $PATH

Stdout:
(empty)

Stderr:
(empty)
"""

Also provides a /stderr handler linked from the status page that
allows all the logs to be viewed.  This is handy when the logs from an
individual test don't provide enough info to debug a build problem.

Change-Id: I4b3ef5a5db6aac1ddf41c41bc2fff310f19a82a1