schema: integrate with the jsonsign package.
Signing schema blobs (permanodes and claims) was too hard before, and
the code was scattered in weird places.

The pkg/jsonsign code tries pretty hard to be Camlistore-neutral, so
it remains so.

The pkg/client code (which deals mostly with HTTP access to a remote
camlistore server) had the signing the logic in it before, mostly
because that's where we originally needed it (for camput), but that
should be fixed in subsequent changes, to remove most of its signing
logic and use pkg/schema's instead.

The schema package now depends on jsonsign, which I think is an
acceptable dependency edge, since the Camlistore schema does logically
depend no signed claims (and that was a gaping hole in its public
API), and it seems safe cyclic-wise, especially if jsonsign continues
to be a leaf package.

Change-Id: I10229ec5d880ddd801cb1ba77ac26b654693036d