Update Closure to ab89cf45c216 (Dec 15, 2013).
I meant to do this in Camlistore commit cc562aff, but that one only
updated to 1389e13 (Jan 23, 2013).

I was confused because the version of third_party/closure/updatelibrary.go
that was checked in prior to cc562aff did not correspond to the version of
Closure that was checked in. So when I ran updatelibrary.go, I got changes
and didn't realize I had to change updatelibrary.go too.

Change-Id: I3ba43a540a62caedbf7570a6d55382bd39a426dc