Reuse memoized mutDirs and mutFiles.
This provides a considerably improved experience for from a single FUSE
client, but not for free.

New dirents introduced outside of FUSE will be visible on the normal
refresh interval.

Removals outside of FUSE are tricky because it's not possible to tell
the difference between a file that was locally created and hasn't made
it into the index yet from a file that was deleted remotely.

Specific dirents that change permanodes outside of FUSE are replaced.

Dirents that change properties outside of FUSE without changing
permanodes (e.g. content, etc...) are never picked up.

Change-Id: I7cc8eaea5951644809053fc97c28455e96508bd4
Fixes:  http://calistore.org/issue/328