Combine the 'Create permanode' feature with the existing 'Create set with x items'.
The web UI was implying 'folderness' by looking for certain claims against
permanodes. This won't allow us to represent empty dynamic folders, which seems
wrong.

Thus I've changed it to just represent all pemanodes as folders. In the future,
if there are more specific types of permanodes, we can handle that.

Change-Id: I73523634114c61f33705a88a4147e13ad60c59d6