Turn Closure's DEBUG flag off.
This improves the performance of Promise construction, which
improves the performance of dropping large numbers of files onto
the UI at once.

There is still more we could do to improve dropping large numbers
of files (probably we should implement a work queue and not try
to start uploading them all in parallel), but this is a quick
solution and probably helps other parts of the code as well.

Change-Id: I398ba8d6776b38b1a61a22598099c52dc2130469