Fix choppy scrolling when flinging on OS X.
This was caused by a bug in Chrome: wheel events stop firing when source
element is removed from DOM. Workaround: leave element in DOM for awhile.

On the upside, React makes this clean to implement. No need to keep state
to remember to remove the dangling element. It automatically gets removed
when it is no longer the last element to receive the wheel event!

Change-Id: I1b9372bfd1edcfb532f6825d99044a314048c3b4