query.go: fix to skip over files returning IsNotExist errors on queries for width, height.
I am a first time committer and I am using the fix to issue number 402 as a bug to learn to learn how to commit to this process. I am not sure if this is the correct way to fix this bug -- the submitter of the bug believes it is -- so please review my code with scrutiny. Thank you.

Change-Id: I6c56b7866d530e1b1c66c563f201e6d1e13d3fa9