Aspects: A new idea for the detail page.
http://imgur.com/Re14XKc,6fYHJqp,39nnQiT#0

Camlistore is built around the idea that every object is a blob,
but blobs can also self-describe themselves as more than a blob.

For example, some blobs are also files, or images, or sets, or
permanodes, or movies, or foursquare checkins. Etc.

Here is an idea for the detail page that reflects that underlying
reality of Camlistore.

Each blob has a single canonical URL in the web UI for its detail
page. Currently this is /?p=<...>, but should ideally be just
/<blobref>.

Within the web UI, many "aspects" register interest in providing
views on blobs. Each time the user navigates to a detail page,
each aspects gets to say whether it can provide a useful view on
the blob.

Aspects are currently rendered (crappily) as tabs along the bottom
of the UI. I'm not sure how they should actually be rendered, this
is temporary.

This patch includes the following aspects:
- image (the old image detail)
- permanode (the old old permanode page)
- blob (the old old blob page)

Change-Id: Idb3cdbb203799a5d9c113d1b37b67a2724040085