pkg/misc/amazon/s3: Fix test broken by commit bf8d6bdb.
Commit bf8d6bdb (s3: auto-select correct region, handle buckets with
dots in hostname) changed the definition of s3.Client replacing the
second field, which was of type *http.Client, with a field of type
http.RoundTripper. But the test utility function getTestClient() was
not modified accordingly. Fix this to make `go test' in
pkg/misc/amazon/s3 succeed again (without this fix the test fails to
build).

Change-Id: Ida383e2050e215cc2446303c0ce59aa7ab6e7981