importers: provide authenticated transport for picasa
And revert context-related changes to importer.go so the other
importers work again.

The problem is picago calls take an *http.Client, so we need to provide
one which already has a transport set up correctly wrt to authentication.
In particular, it needs to have the access token that was obtained
during oauth.

Change-Id: I7b83cd80910111a9abf1c192cc59841f6d818fb7