picasa importer: remove Run-private HTTPClient
A more complete version of Aaron Boodman's
https://camlistore-review.googlesource.com/#/c/3066/
in removing Transports. As Aaron pointed out,
(https://groups.google.com/d/msg/camlistore/9OwQKH6z7rY/840oQrRgsGIJ) it
is illegal for oauth.Transport.Transport to ever be another instance of
oauth.Transport.

So this commit tries hard to use Context's HTTPClient
everywhere and make sure to have its Transport an *oauth.Transport, but
that's Transport is not an *oauth.Transport, but the underlying
Transport.

To eliminate "403: Token not found" errors, we set
oauth.Config.AccessType, oauth.Config.ApprovalPrompt to
"offline", "force".
This rewards us with a non-empty Token.RefreshToken,
and store the full Token in AcctAttrOAuthToken.
This enables us to import continuously without user intervention.

Change-Id: I4defb346abfcf3f6c17597fc2dc23a6e536eda37