ui/permanode: cope better with not a permanode case
A camliPath can now have a file blob as a target (instead of a
permanode). And when we render these camliPath members as links, since
it can happen that these members haven't been described, we can't know
in advance whether to link them as permanodes or as blobs, so we
always default to permanode. Therefore, it can now legitimately happen
that we end up on the permanode page, while the blobRef is actually
one of a non permanode.

This patch tries to cope with that situation in a bit nicer way.

Context: http://camlistore.org/issue/470

Change-Id: Ic3b728541e164a6e6056d8f275a1d7e76b41e043