Checking in an unmodified copy of Go's archive/zip package.
The import is happening at hg 47de34bbcca1 but the last modification
of the package was at:

changeset:   20518:8ab50c14cd95
user:        Brad Fitzpatrick <bradfitz@golang.org>
date:        Mon Aug 04 16:12:55 2014 -0700
summary:     archive/zip: accept bogus trailing zeros in extras

... which will go into Go 1.4.

We don't need that fix in Camlistore, but rather we need additional
data about the state of the *zip.Writer while it's writing.  So this
is a pristine import in preparation for future changes so the changes
will be easy to see in the history, in case we want to move the
changes to the Go standard library (likely). But Go goes into 1.4
freeze in two days and it's the weekend.

Change-Id: I07117b7c5a7f58bd1e8abf6227c5fb0fa8047fa3