Fix FOPA (flash of permanode aspect) when clicking permanodes with children.
We were waiting to show the container aspect until the child search session
returned. This meant that we temporarily showed the permanode aspect then
swapped to the container aspect shortly thereafter (because the container
aspect is higher priority).

The right solution to this is to change the search protocol so that clients
can know how many children a permanode has without getting the entire list
of them. But for now, it just so happens that search returns the entire
list, so we can use that instead.

Change-Id: Iaf93f689e965a1dc9ef73d834d46c57026c100cd