pkg/client: make Describe use a POST request to the server
Otherwise it can't send Describe Rules.

Also in this commit: a bunch of new tests from when I thought I was going crazy
and trying to debug the search system, before I realized the problem was that the
client was doing a GET request and dropping most of my DescribeRequest.

Change-Id: I4ea9bed80f0d7d6b86814527b63acc3586ac1d06