pkg/fs: implement magic folder "versions" in fuse
A file is shown as folder and inside you can see all its versions (named according
to the date). Basically is like `at` but the path (not the date) goes first.

So if you go to:

    $ ls  <mountpoint>/versions/my_folder
    dr-x------  1 grecco  32000  0 Mar 13 01:39 my_file

and then:

    $ ls  <mountpoint>/versions/my_folder/my_file
    -r--------  1 grecco  32000   2 Mar 13 01:39 2014-03-12T14:53:34.471588505Z
    -r--------  1 grecco  32000   2 Mar 13 01:39 2014-03-12T14:53:36.6569929Z
    -r--------  1 grecco  32000   2 Mar 13 01:39 2014-03-12T14:53:38.842875168Z
    -r--------  1 grecco  32000  25 Mar 13 01:39 2014-03-12T21:16:09.905612807Z

These files are standard files which can be opened to see the file content as a
specific point of time.

Change-Id: I38a4d7bf35ba32407036535e629039e23dc32735