misc/buildbot: keep it alive until rewrite
Because of the hg->git switch and the switch to Go 1.4, which requires
bootstrapping for building from source:

-master now requires an env with a working Go to build builder (instead of
building go tip from source), and master does not monitor changes to the go
tree anymore.

-builder gets Go1.4.1 from the binary tarball (instead of building from
source), and does not build go tip anymore (nor does it run tests with
it).

Keeping it alive properly was not deemed worth the trouble since we plan
on replacing it with a new builder based on the same tools used for the
Go builders.

Change-Id: Idd0864350b27b6523675006f00719fba2d6fc72f