index/sqlite tests: let sorted/sqlite pkg do initDB and enable WAL
The problem was newSorted was taking care of doing the DB
initialization. It was not only unnecessary, since newKeyValueFromConfig
from sorted/sqlite will do it by calling initDB, but also incorrect
because WAL was not getting enabled. Which is what was causing the
"database is locked" error "fixed" in commit
f6bf779cd6a8eda505c8392cbbba9d9d0fc4c96e

Change-Id: Ib4badeda4ed4c6c4be43d5f7fc4c543a467b0cd1